+++ test/test_Pipeline.py
@@ -1,16 +1,22 @@
 from scalecast.Pipeline import Pipeline, MVPipeline, Transformer, Reverter
-from scalecast.util import find_statistical_transformation, find_optimal_transformation
-from scalecast.util import break_mv_forecaster
+from scalecast.util import (
+    find_statistical_transformation, 
+    find_optimal_transformation, 
+    break_mv_forecaster,
+    backtest_metrics,
+) 
 from test_SeriesTransformer import forecaster
 from test_Forecaster import build_Forecaster
 from test_MVForecaster import build_MVForecaster
 
 def mv_forecaster(mvf):
     mvf.set_estimator('elasticnet')
-    mvf.manual_forecast(lags=12)
+    mvf.manual_forecast(lags=12,alpha=.2)
+    mvf.set_estimator('ridge')
+    mvf.manual_forecast(lags=12,alpha=.2)
 
 def test_pipeline():
-    f = build_Forecaster()
+    f = build_Forecaster(cis=True)
     transformer, reverter = find_optimal_transformation(
         f,
         estimator='elasticnet',
@@ -23,6 +29,9 @@ def test_pipeline():
         ],
     )
     f = pipeline.fit_predict(f)
+    backtest_results = pipeline.backtest(f)
+    backtest_mets = backtest_metrics(backtest_results)
+    backtest_mets.to_excel('../../uv_backtest_results.xlsx')
 
 def test_mvpipeline():
     mvf = build_MVForecaster()
@@ -45,9 +54,20 @@ def test_mvpipeline():
             ('Forecast',mv_forecaster),
             ('Revert',[reverter1,reverter2,reverter3]),
         ],
-        test_length = 12,
+        test_length = 20,
+        cis = True,
     )
     f1, f2, f3 = pipeline.fit_predict(f1,f2,f3)
+    backtest_results = pipeline.backtest(f1,f2,f3,n_iter=2,jump_back=12)
+    backtest_mets = backtest_metrics(
+        backtest_results,
+        mets = ['rmse','smape','mape','r2','mae'],
+        names=['UTUR','UNRATE','SAHMREALTIME'],
+        mase = True,
+        msis = True,
+        m = 12,
+    )
+    backtest_mets.to_excel('../../mv_backtest_results.xlsx')
 
 def main():
     test_pipeline()